+++ src/main/java/de/maxhenkel/plane/ServerConfig.java
@@ -3,10 +3,9 @@ package de.maxhenkel.plane;
 import de.maxhenkel.corelib.config.ConfigBase;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.material.Fluid;
-import net.minecraftforge.common.ForgeConfigSpec;
-import net.minecraftforge.fml.event.config.ModConfigEvent;
-import net.minecraftforge.registries.ForgeRegistries;
-
+import net.neoforged.fml.event.config.ModConfigEvent;
+import net.neoforged.neoforge.common.ModConfigSpec;
+import net.neoforged.neoforge.registries.ForgeRegistries;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -15,10 +14,10 @@ import java.util.stream.Collectors;
 
 public class ServerConfig extends ConfigBase {
 
-    private final ForgeConfigSpec.ConfigValue<List<? extends String>> validFuelsSpec;
+    private final ModConfigSpec.ConfigValue<List<? extends String>> validFuelsSpec;
     public static List<Fluid> validFuels = new ArrayList<>();
 
-    public ServerConfig(ForgeConfigSpec.Builder builder) {
+    public ServerConfig(ModConfigSpec.Builder builder) {
         super(builder);
         validFuelsSpec = builder.defineList("valid_fuels", Arrays.asList("car:bio_diesel"), Objects::nonNull);
     }