+++ src/main/java/de/maxhenkel/plane/entity/EntityCargoPlane.java
@@ -24,6 +24,7 @@ import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.ChestMenu;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
+import net.neoforged.neoforge.network.NetworkHooks;
 
 import javax.annotation.Nullable;
 
@@ -70,7 +71,7 @@ public class EntityCargoPlane extends EntityPlaneSoundBase {
                     }
                 });
             } else {
-                serverPlayer.openMenu(new MenuProvider() {
+                NetworkHooks.openScreen(serverPlayer, new MenuProvider() {
                     @Override
                     public Component getDisplayName() {
                         return getName();