+++ src/main/java/de/maxhenkel/plane/events/CreativeTabEvents.java
@@ -6,9 +6,9 @@ import de.maxhenkel.plane.item.ItemPlane;
 import de.maxhenkel.plane.item.ModItems;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.registries.RegistryObject;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
+import net.neoforged.neoforge.registries.RegistryObject;
 
 public class CreativeTabEvents {